requirements.txt
flask==2.1.3
Werkzeug==2.2.2
openai==0.28.0
flask-cors==4.0.0
pytest==7.4.3
python-dotenv==1.0.0